Make SentryOptions inherit from SentryCommonOptions and use reflection to copy values.
